+++ src/main/java/com/hanghae/newsfeed/admin/controller/AdminPostController.java
@@ -1,11 +1,11 @@
 package com.hanghae.newsfeed.admin.controller;
 
-import com.hanghae.newsfeed.admin.service.AdminPostService;
+import com.hanghae.newsfeed.admin.service.impl.AdminPostServiceImpl;
 import com.hanghae.newsfeed.post.dto.request.PostRequest;
 import com.hanghae.newsfeed.post.dto.response.MultimediaResponse;
 import com.hanghae.newsfeed.post.dto.response.PostResponse;
-import com.hanghae.newsfeed.post.service.MultimediaService;
-import com.hanghae.newsfeed.post.service.PostService;
+import com.hanghae.newsfeed.post.service.impl.MultimediaServiceImpl;
+import com.hanghae.newsfeed.post.service.impl.PostServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +17,9 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/admins/posts")
 public class AdminPostController {
-    private final PostService postService;
-    private final AdminPostService adminPostService;
-    private final MultimediaService multimediaService;
+    private final PostServiceImpl postService;
+    private final AdminPostServiceImpl adminPostService;
+    private final MultimediaServiceImpl multimediaService;
 
     // 게시물 목록 조회
     @GetMapping
