+++ src/main/java/com/hanghae/newsfeed/admin/controller/AdminUserController.java
@@ -2,7 +2,7 @@ package com.hanghae.newsfeed.admin.controller;
 
 import com.hanghae.newsfeed.admin.dto.request.AdminUserRequest;
 import com.hanghae.newsfeed.admin.dto.response.AdminUserResponse;
-import com.hanghae.newsfeed.admin.service.AdminUserService;
+import com.hanghae.newsfeed.admin.service.impl.AdminUserServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +14,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/admins")
 public class AdminUserController {
-    private final AdminUserService adminUserService;
+    private final AdminUserServiceImpl adminUserService;
 
     // 회원 목록 조회
     @GetMapping("/users")
