+++ src/main/java/com/hanghae/newsfeed/admin/service/AdminCommentService.java
@@ -2,46 +2,20 @@ package com.hanghae.newsfeed.admin.service;
 
 import com.hanghae.newsfeed.comment.dto.request.CommentRequest;
 import com.hanghae.newsfeed.comment.dto.response.CommentResponse;
-import com.hanghae.newsfeed.comment.entity.Comment;
-import com.hanghae.newsfeed.comment.repository.CommentRepository;
-import com.hanghae.newsfeed.common.exception.CustomErrorCode;
-import com.hanghae.newsfeed.common.exception.CustomException;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class AdminCommentService {
-    private final CommentRepository commentRepository;
+public interface AdminCommentService {
+    /**
+     * 댓글 수정
+     * @param commentId 댓글 아이디
+     * @param request 댓글 수정 요청
+     * @return 수정된 댓글 정보
+     */
+    CommentResponse updateComment(Long commentId, CommentRequest request);
 
-    // 댓글 수정
-    @Transactional
-    public CommentResponse updateComment(Long commentId, CommentRequest request) {
-        // 댓글 조회 예외 발생
-        Comment target = commentRepository.findById(commentId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.COMMENT_NOT_FOUND));
-
-        // 댓글 수정
-        target.updateComment(request);
-
-        // DB로 갱신
-        Comment updatedComment = commentRepository.save(target);
-
-        return CommentResponse.createCommentDto(updatedComment, "댓글 수정 성공");
-    }
-
-    // 댓글 삭제
-    @Transactional
-    public CommentResponse deleteComment(Long commentId) {
-        // 댓글 조회 예외 발생
-        Comment target = commentRepository.findById(commentId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.COMMENT_NOT_FOUND));
-
-        // 댓글 삭제
-        commentRepository.delete(target);
-
-        return CommentResponse.createCommentDto(target, "댓글 삭제 성공");
-    }
-
+    /**
+     * 댓글 삭제
+     * @param commentId 댓글 아이디
+     * @return 삭제된 댓글 정보
+     */
+    CommentResponse deleteComment(Long commentId);
 }
