+++ src/main/java/com/hanghae/newsfeed/admin/service/AdminPostService.java
@@ -1,46 +1,21 @@
 package com.hanghae.newsfeed.admin.service;
 
-import com.hanghae.newsfeed.common.exception.CustomErrorCode;
-import com.hanghae.newsfeed.common.exception.CustomException;
 import com.hanghae.newsfeed.post.dto.request.PostRequest;
 import com.hanghae.newsfeed.post.dto.response.PostResponse;
-import com.hanghae.newsfeed.post.entity.Post;
-import com.hanghae.newsfeed.post.repository.PostRepository;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class AdminPostService {
-    private final PostRepository postRepository;
+public interface AdminPostService {
+    /**
+     * 게시물 수정
+     * @param postId 게시물 아이디
+     * @param request 게시물 수정 요청
+     * @return 수정된 게시물 정보
+     */
+    PostResponse updatePost(Long postId, PostRequest request);
 
-    // 게시물 수정
-    @Transactional
-    public PostResponse updatePost(Long postId, PostRequest request) {
-        // 게시물 조회 예외 발생
-        Post target = postRepository.findById(postId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.POST_NOT_FOUND));
-
-        // 게시물 수정
-        target.updatePost(request);
-
-        // DB로 갱신
-        Post updatedPost = postRepository.save(target);
-
-        return PostResponse.createPostDto(updatedPost, "게시물 수정 성공");
-    }
-
-    // 게시물 삭제
-    @Transactional
-    public PostResponse deletePost(Long postId) {
-        // 게시물 조회 예외 발생
-        Post target = postRepository.findById(postId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.POST_NOT_FOUND));
-
-        // 게시물 삭제
-        postRepository.delete(target);
-
-        return PostResponse.createPostDto(target, "게시물 삭제 성공");
-    }
+    /**
+     * 게시물 삭제
+     * @param postId 게시물 아이디
+     * @return 삭제된 게시물 정보
+     */
+    PostResponse deletePost(Long postId);
 }
