+++ src/main/java/com/hanghae/newsfeed/admin/service/AdminUserService.java
@@ -2,54 +2,21 @@ package com.hanghae.newsfeed.admin.service;
 
 import com.hanghae.newsfeed.admin.dto.request.AdminUserRequest;
 import com.hanghae.newsfeed.admin.dto.response.AdminUserResponse;
-import com.hanghae.newsfeed.common.exception.CustomErrorCode;
-import com.hanghae.newsfeed.common.exception.CustomException;
-import com.hanghae.newsfeed.user.entity.User;
-import com.hanghae.newsfeed.user.repository.UserRepository;
-import com.hanghae.newsfeed.user.type.UserRoleEnum;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-public class AdminUserService {
-    private final UserRepository userRepository;
+public interface AdminUserService {
+    /**
+     * 전체 회원 목록 조회
+     * @return 회원 목록
+     */
+    List<AdminUserResponse> getAllUsers();
 
-    // 전체 회원 목록 조회
-    public List<AdminUserResponse> getAllUsers() {
-        List<User> allUsers = userRepository.findAll();
-
-        return allUsers.stream()
-                .map(user -> AdminUserResponse.createAdminUserDto(user, "전체 회원 조회 성공"))
-                .collect(Collectors.toList());
-    }
-
-    // 회원 권한 수정(USER -> ADMIN / active = false)
-    public AdminUserResponse updateUserRoleAndStatus(Long userId, AdminUserRequest request) {
-        // 유저 조회 예외 발생
-        User target = userRepository.findById(userId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.USER_NOT_FOUND));
-
-        // 이미 탈퇴한 회원은 수정 불가능
-        if (!target.getActive()) {
-            throw new CustomException(CustomErrorCode.USER_DEACTIVATED);
-        }
-
-        // 관리자의 role과 active 수정 불가능
-        if (target.getRole() == UserRoleEnum.ADMIN) {
-            throw new CustomException(CustomErrorCode.ADMIN_CANNOT_BE_MODIFIED);
-        }
-
-        // 유저 수정
-        target.updateUserRoleAndActive(request);
-
-        // DB로 갱신
-        User updatedUser = userRepository.save(target);
-
-        // DTO로 변경하여 반환
-        return AdminUserResponse.createAdminUserDto(updatedUser, "회원 권한 수정 성공");
-    }
+    /**
+     * 회원 권한 수정
+     * @param userId 사용자 아이디
+     * @param request 회원 권한 수정 요청
+     * @return 수정된 회원 정보
+     */
+    AdminUserResponse updateUserRoleAndStatus(Long userId, AdminUserRequest request);
 }
