+++ src/main/java/com/hanghae/newsfeed/admin/service/impl/AdminCommentServiceImpl.java
@@ -1,0 +1,50 @@
+package com.hanghae.newsfeed.admin.service.impl;
+
+import com.hanghae.newsfeed.admin.service.AdminCommentService;
+import com.hanghae.newsfeed.comment.dto.request.CommentRequest;
+import com.hanghae.newsfeed.comment.dto.response.CommentResponse;
+import com.hanghae.newsfeed.comment.entity.Comment;
+import com.hanghae.newsfeed.comment.repository.CommentRepository;
+import com.hanghae.newsfeed.common.exception.CustomErrorCode;
+import com.hanghae.newsfeed.common.exception.CustomException;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AdminCommentServiceImpl implements AdminCommentService {
+    private final CommentRepository commentRepository;
+
+    // 댓글 수정
+    @Override
+    @Transactional
+    public CommentResponse updateComment(Long commentId, CommentRequest request) {
+        // 댓글 조회 예외 발생
+        Comment target = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CustomException(CustomErrorCode.COMMENT_NOT_FOUND));
+
+        // 댓글 수정
+        target.updateComment(request);
+
+        // DB로 갱신
+        Comment updatedComment = commentRepository.save(target);
+
+        return CommentResponse.createCommentDto(updatedComment, "댓글 수정 성공");
+    }
+
+    // 댓글 삭제
+    @Override
+    @Transactional
+    public CommentResponse deleteComment(Long commentId) {
+        // 댓글 조회 예외 발생
+        Comment target = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CustomException(CustomErrorCode.COMMENT_NOT_FOUND));
+
+        // 댓글 삭제
+        commentRepository.delete(target);
+
+        return CommentResponse.createCommentDto(target, "댓글 삭제 성공");
+    }
+
+}
