+++ src/main/java/com/hanghae/newsfeed/auth/controller/AuthController.java
@@ -5,7 +5,7 @@ import com.hanghae.newsfeed.auth.dto.request.SignupRequest;
 import com.hanghae.newsfeed.auth.dto.response.LoginResponse;
 import com.hanghae.newsfeed.auth.dto.response.LogoutResponse;
 import com.hanghae.newsfeed.auth.dto.response.SignupResponse;
-import com.hanghae.newsfeed.auth.service.AuthService;
+import com.hanghae.newsfeed.auth.service.impl.AuthServiceImpl;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/api/auth")
 public class AuthController {
-    private final AuthService authService;
+    private final AuthServiceImpl authService;
 
     // 회원 가입
     @PostMapping("/signup")
