+++ src/main/java/com/hanghae/newsfeed/like/controller/CommentLikeController.java
@@ -1,7 +1,7 @@
 package com.hanghae.newsfeed.like.controller;
 
 import com.hanghae.newsfeed.like.dto.response.CommentLikeResponse;
-import com.hanghae.newsfeed.like.service.CommentLikeService;
+import com.hanghae.newsfeed.like.service.impl.CommentLikeServiceImpl;
 import com.hanghae.newsfeed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/api/likes")
 public class CommentLikeController {
-    private final CommentLikeService commentLikeService;
+    private final CommentLikeServiceImpl commentLikeService;
 
     // 댓글 좋아요
     @PostMapping("/comments/{commentId}")
