+++ src/main/java/com/hanghae/newsfeed/like/controller/PostLikeController.java
@@ -1,7 +1,7 @@
 package com.hanghae.newsfeed.like.controller;
 
 import com.hanghae.newsfeed.like.dto.response.PostLikeResponse;
-import com.hanghae.newsfeed.like.service.PostLikeService;
+import com.hanghae.newsfeed.like.service.impl.PostLikeServiceImpl;
 import com.hanghae.newsfeed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/api/likes")
 public class PostLikeController {
-    private final PostLikeService postLikeService;
+    private final PostLikeServiceImpl postLikeService;
 
     // 게시물 좋아요
     @PostMapping("/posts/{postId}")
