+++ src/main/java/com/hanghae/newsfeed/like/service/PostLikeService.java
@@ -1,70 +1,22 @@
 package com.hanghae.newsfeed.like.service;
 
-import com.hanghae.newsfeed.common.exception.CustomErrorCode;
-import com.hanghae.newsfeed.common.exception.CustomException;
-import com.hanghae.newsfeed.like.dto.response.PostLikeResponse;
-import com.hanghae.newsfeed.like.entity.PostLike;
-import com.hanghae.newsfeed.like.repository.PostLikeRepository;
-import com.hanghae.newsfeed.post.entity.Post;
-import com.hanghae.newsfeed.post.repository.PostRepository;
 import com.hanghae.newsfeed.auth.security.UserDetailsImpl;
-import com.hanghae.newsfeed.user.entity.User;
-import com.hanghae.newsfeed.user.repository.UserRepository;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+import com.hanghae.newsfeed.like.dto.response.PostLikeResponse;
 
-@Service
-@RequiredArgsConstructor
-public class PostLikeService {
-    private final PostLikeRepository postLikeRepository;
-    private final PostRepository postRepository;
-    private final UserRepository userRepository;
+public interface PostLikeService {
+    /**
+     * 게시물 좋아요
+     * @param userDetails 사용자 상세 정보
+     * @param postId 게시물 아이디
+     * @return 좋아요 결과
+     */
+    PostLikeResponse likePost(UserDetailsImpl userDetails, Long postId);
 
-    // 게시물 좋아요
-    @Transactional
-    public PostLikeResponse likePost(UserDetailsImpl userDetails, Long postId) {
-        User user = userRepository.findById(userDetails.getId())
-                .orElseThrow(() -> new CustomException(CustomErrorCode.USER_NOT_FOUND));
-
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.POST_NOT_FOUND));
-
-        // 작성자 본인이 아닌지 확인
-        if (!user.equals(post.getUser())) {
-            // 기존 좋아요 여부 확인
-            if (postLikeRepository.existsByUserAndPost(user, post)) {
-                throw new CustomException(CustomErrorCode.ALREADY_LIKED);
-            }
-
-            // 좋아요 등록
-            PostLike postLike = new PostLike(user, post);
-            postLikeRepository.save(postLike);
-
-            return new PostLikeResponse(user.getId(), post.getId(), "게시물 좋아요 성공");
-        } else {
-            throw new CustomException(CustomErrorCode.CANNOT_LIKE_OWN_CONTENT);
-        }
-    }
-
-    // 게시물 좋아요 취소
-    @Transactional
-    public PostLikeResponse unlikePost(UserDetailsImpl userDetails, Long postId) {
-        User user = userRepository.findById(userDetails.getId())
-                .orElseThrow(() -> new CustomException(CustomErrorCode.USER_NOT_FOUND));
-
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.POST_NOT_FOUND));
-
-        // 기존 좋아요 여부 확인
-        if (postLikeRepository.existsByUserAndPost(user, post)) {
-            PostLike postLike = postLikeRepository.findByUserAndPost(user, post);
-
-            postLikeRepository.delete(postLike);
-
-            return new PostLikeResponse(user.getId(), post.getId(), "게시물 좋아요 취소 성공");
-        } else {
-            throw new CustomException(CustomErrorCode.NO_LIKE_YET);
-        }
-    }
+    /**
+     * 게시물 좋아요 취소
+     * @param userDetails 사용자 상세 정보
+     * @param postId 게시물 아이디
+     * @return 좋아요 취소 결과
+     */
+    PostLikeResponse unlikePost(UserDetailsImpl userDetails, Long postId);
 }
